requirements.txt
pyserial==3.5
fastmcp==2.5.2
fastapi==0.115.12
fastapi-cli==0.0.7
requests==2.32.3